where_to_go/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-admin-sortable2==2.0
django-environ==0.8.1
django-tinymce==3.4.0
environs==9.5.0
marshmallow==3.15.0
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.7
python-dotenv==0.20.0
sqlparse==0.4.2
